fix(bootstrap): map singleton_id bang Byte cho khop cot TINYINT
Entity khai bao Integer trong khi V10 tao cot TINYINT. Tren Aiven,
ddl-auto=validate tu choi tao entityManagerFactory nen ung dung khong
khoi dong duoc va runner bootstrap admin khong bao gio chay.

V10 da apply tren database dang chay nen sua entity, khong sua migration.

Bo sung SchemaValidationIT doi chieu truc tiep information_schema voi
metamodel. Da do thuc nghiem: tren MySQL 8.0 cua Testcontainers,
ddl-auto=validate bat duoc cot thieu nhung khong bat lech do rong kieu
so nguyen (ca Integer lan Long deu qua), nen khong the dua vao rieng
validate de chan lop loi nay.

+++ src/main/java/com/example/book_be/nguoidung/domain/AdminBootstrapState.java
@@ -15,11 +15,15 @@ import java.util.Date;
 @Table(name = "admin_bootstrap_state")
 public class AdminBootstrapState {
 
-    public static final int SINGLETON_ID = 1;
+    public static final byte SINGLETON_ID = 1;
 
+    /**
+     * Cot la TINYINT trong V10. Entity phai khai bao Byte cho khop, neu khong
+     * ddl-auto=validate se tu choi khoi dong ung dung.
+     */
     @Id
     @Column(name = "singleton_id")
-    private Integer singletonId;
+    private Byte singletonId;
 
     @Column(name = "da_su_dung", nullable = false)
     private boolean daSuDung;
@@ -33,7 +37,7 @@ public class AdminBootstrapState {
     protected AdminBootstrapState() {
     }
 
-    public Integer getSingletonId() {
+    public Byte getSingletonId() {
         return singletonId;
     }
